% Generated by roxygen2 (4.0.2): do not edit by hand
\name{hello}
\alias{hello}
\title{x3dom hello world example}
\usage{
hello(width = NULL, height = NULL)
}
\arguments{
\item{width}{The container div width.}

\item{height}{The container div height.}
}
\description{
Display three simple objects in the x3dom.
}

