% Generated by roxygen2 (4.0.2): do not edit by hand
\name{scatterplotx3}
\alias{scatterplotx3}
\title{scatterplotx3 x3dom 3D scatterplot example}
\source{
Harry Voorhees: \url{http://bl.ocks.org/hlvoorhees/5986172}.
}
\usage{
scatterplotx3(x, height = NULL, width = NULL, axis = TRUE,
  tick.marks = TRUE, num.ticks = c(6, 6, 6), color = "steelblue",
  grid = TRUE)
}
\arguments{
\item{x}{A data matrix with three columns corresponding to the x,y,z
coordinate axes. Column labels, if present, are used as axis labels.}

\item{height}{The container div height.}

\item{width}{The container div width.}

\item{tick.marks}{Set FALSE to disable display of tick marks.}

\item{num.ticks}{A three-element vector with the suggested number of
ticks to display per axis.}

\item{color}{Either a single valid RGB, hex or named color name, or
a vector of valid color names of length \code{nrow(x)}.}

\item{grid}{Set FALSE to disable display of a grid.}
}
\description{
A basic 3D scatterplot using d3.js and x3dom.
}
\examples{
## dontrun
# A stand-alone example
set.seed(1)
x <- matrix(rnorm(100*3),ncol=3)
scatterplotx3(x)

# A shiny example
runApp(system.file("examples/shiny",package="x3dom"))
}

